Updated src/Image_lib.rs
