+++ src/main.rs
@@ -1,16 +1,4 @@
-use image::imageops::FilterType;
-use std::path::Path;
-use std::env;
-
-fn main() {
-        let args: Vec<String> = env::args().collect();
-        if args.len() < 2 {
-                panic!("not enough arguments please specify the file name");
-        }
-        let filename = &args[1]; 
-        let img = image::open(filename).unwrap();
-        let scaled = img.resize(299, 299, FilterType::CatmullRom); 
-        let path = Path::new("image_resized.jpg");
-        scaled.save(path).unwrap();
+fn main(){
+        println!("hello world");
 }
